refactor(core): improvements
