+++ src/Worker/ExecutionPolicy/FiberPolicy.php
@@ -5,8 +5,7 @@ declare(strict_types=1);
 namespace SchedulerBundle\Worker\ExecutionPolicy;
 
 use Closure;
-use SchedulerBundle\Task\TaskListInterface;
-use SchedulerBundle\Worker\WorkerConfiguration;
+use SchedulerBundle\Task\TaskInterface;
 use SchedulerBundle\Worker\WorkerInterface;
 
 /**
@@ -17,9 +16,34 @@ final class FiberPolicy implements ExecutionPolicyInterface
     /**
      * {@inheritdoc}
      */
-    public function execute(WorkerConfiguration $configuration, TaskListInterface $taskList): Closure
-    {
-        return static function (WorkerInterface $worker, TaskListInterface $taskList): void {};
+    public function execute(
+        WorkerInterface $worker,
+        Closure $fetchTaskListFunc,
+        Closure $handleTaskFunc
+    ): void {
+        while (!$worker->getConfiguration()->shouldStop()) {
+            $toExecuteTasks = $fetchTaskListFunc();
+            if (0 === $toExecuteTasks->count() && !$worker->getConfiguration()->isSleepingUntilNextMinute()) {
+                $worker->stop();
+            }
+
+            $toExecuteTasks->walk(function (TaskInterface $task) use ($toExecuteTasks, $handleTaskFunc): void {
+                $fiber = new Fiber(function (TaskInterface $toExecuteTask) use ($toExecuteTasks, $handleTaskFunc): void {
+                    $handleTaskFunc($toExecuteTask, $toExecuteTasks);
+                });
+
+                $fiber->start($task, $toExecuteTasks);
+            });
+
+            if ($worker->getConfiguration()->shouldStop()) {
+                break;
+            }
+
+            if ($worker->getConfiguration()->isSleepingUntilNextMinute()) {
+                $worker->sleep();
+                $this->execute($worker, $fetchTaskListFunc, $handleTaskFunc);
+            }
+        }
     }
 
     /**
